Fix Kotlin type mismatch in LiteRtLmPlugin SamplerConfig
LiteRT-LM SamplerConfig expects Double, not Float for topP and temperature.

+++ android/app/src/main/kotlin/com/appforge/pocketai/LiteRtLmPlugin.kt
@@ -125,8 +125,8 @@ class LiteRtLmPlugin private constructor(
                     val conv = engine!!.createConversation(ConversationConfig(
                         samplerConfig = SamplerConfig(
                             topK = topK,
-                            topP = topP.toFloat(),
-                            temperature = temperature.toFloat()
+                            topP = topP,
+                            temperature = temperature
                         )
                     ))
                     conversation = conv
